random windows fixes and updates
git-svn-id: https://commonmusic.svn.sourceforge.net/svnroot/commonmusic/trunk@1646 d60aafaf-7936-0410-ba4d-d307febf7868

+++ grace/src/Midi.cpp
@@ -14,7 +14,7 @@
 #include "Toolbox.h"
 #include "Editor.h"
 
-#include <iostream>;
+#include <iostream>
 
 //
 //  Nodes
@@ -2172,7 +2172,7 @@ void MidiFileInfoComponent::buttonClicked (Button* button)
 	    return;
 	}
       midifileinfo->file=fileeditor->getCurrentFile();
-      midifileinfo->tempo=temposlider->getValue();
+      midifileinfo->tempo=(int)temposlider->getValue();
       midifileinfo->keysig=keysigmenu->getSelectedId();
       midifileinfo->insts=tuningbutton->getToggleState();
       midifileinfo->bends=instrumentbutton->getToggleState();
